requirements.txt
Flask==2.0.1
openai==0.27.0
pdfminer.six==20201018
